requirements.txt
bleak
Pillow
requests
